+++ packages/plugin/src/controllers/SubmissionsController.php
@@ -195,7 +195,7 @@ class SubmissionsController extends BaseController
             'selectedSite' => $selectedSite,
         ];
 
-        $variables = array_merge($variables, $craftVersionSpecificVariables);
+        $variables = array_merge($variables, $craftVersionSpecificVariables, $this->getEditTemplateVariables($submission));
 
         return $this->renderTemplate(
             $this->getTemplateBasePath().'/edit',
@@ -329,6 +329,11 @@ class SubmissionsController extends BaseController
     protected function getTemplateBasePath(): string
     {
         return self::TEMPLATE_BASE_PATH;
+    }
+
+    protected function getEditTemplateVariables(Submission $submission): array
+    {
+        return [];
     }
 
     private function checkPermissions(Form $form): void